requirements.txt

aiohttp>=3.8.0
feedparser>=6.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tweepy>=4.14.0
langdetect>=1.0.9
nltk>=3.8.0
textblob>=0.17.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced functionality
beautifulsoup4>=4.12.0
requests>=2.31.0
